+++ tests/testthat/test-createDSR.R
@@ -137,4 +137,3 @@
 #   unlink(x = file.path(pathstd, "test_case"), recursive = TRUE)
 #   
 # })
-