+++ tests/testthat/test-createPSP.R
@@ -175,4 +175,3 @@
 #   unlink(x = file.path(pathstd, "test_case"), recursive = TRUE)
 #   
 # })
-